Add class for managing session, moved all entities into package 'entity'
