+++ model/src/main/java/com/college/model/entity/Automobile.java
@@ -2,11 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.college.model;
+package com.college.model.models;
 
 import jakarta.persistence.*;
-
-import java.util.Map;
 
 /**
  * @author Vagid Zibliuc
