+++ model/src/main/java/com/college/model/entity/SimpleTable.java
@@ -1,4 +1,4 @@
-package com.college.model;
+package com.college.model.models;
 
 public abstract class SimpleTable {
