+++ model/src/main/java/com/college/model/entity/Warranty.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.college.model;
+package com.college.model.models;
 
 import jakarta.persistence.*;
