+++ model/src/test/java/com/college/model/database/implementations/UserImplTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class UserImplTest {
+  
+}
